% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ippuc_neigh}
\alias{ippuc_neigh}
\title{Neighborhood Boundaries Data for Curitiba}
\format{
A data frame and \code{sf} object with 75 observations and 10 variables:
\describe{
\item{OBJECTID}{num. The unique identifier for each neighborhood boundary object.}
\item{CODIGO}{num. The code representing the neighborhood.}
\item{TIPO}{chr. The type of boundary, here it is "DIVISA DE BAIRROS".}
\item{NOME}{chr. The name of the neighborhood.}
\item{FONTE}{chr. The source of the boundary information, usually a municipal decree.}
\item{CD_REGIONA}{num. The code representing the region to which the neighborhood belongs.}
\item{NM_REGIONA}{chr. The name of the region to which the neighborhood belongs.}
\item{SHAPE_AREA}{num. The area of the neighborhood in square meters.}
\item{SHAPE_LEN}{num. The perimeter length of the neighborhood in meters.}
\item{geometry}{sfc_POLYGON. The spatial geometry of the neighborhood boundary.}
}
}
\source{
Curitiba Municipal Government (IPPUC).
}
\usage{
ippuc_neigh
}
\description{
A dataset containing the neighborhood boundaries for Curitiba, Brazil.
}
\keyword{datasets}
