% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ippuc_road_axis}
\alias{ippuc_road_axis}
\title{Road Data for Curitiba}
\format{
A data frame and \code{sf} object with 40,355 observations and 22 variables:
\describe{
\item{CHAVE}{chr. The unique key identifier for each road segment.}
\item{CODVIA}{chr. The code of the road.}
\item{NOINICIO}{chr. The starting number of the road segment.}
\item{NOFIM}{chr. The ending number of the road segment.}
\item{STATUS}{chr. The status of the road segment (e.g., "OFICIAL").}
\item{NMVIA}{chr. The name of the road.}
\item{NMVIA_ANT}{chr. The former name of the road, if any.}
\item{SVIARIO}{chr. The classification of the road.}
\item{SVIARIOLEG}{chr. The legal reference for the road classification.}
\item{HIERARQUIA}{chr. The hierarchy level of the road.}
\item{STATUS_DEM}{chr. The official status of the road.}
\item{COD_BAIRRO}{num. The code of the neighborhood}
\item{BAIRRO_E}{chr. The name of the neighborhood on the left side of the road.}
\item{BAIRRO_D}{chr. The name of the neighborhood on the right side of the road.}
\item{COD_REG}{num. The code of the region.}
\item{REGIONAL_E}{chr. The name of the regional administration on the left side of the road.}
\item{REGIONAL_D}{chr. The name of the regional administration on the right side of the road.}
\item{CEP_E}{chr. The postal code on the left side of the road.}
\item{CEP_D}{chr. The postal code on the right side of the road.}
\item{CEP_INFO}{chr. Information about the postal code.}
\item{CEP_FONTE}{chr. The source of the postal code information.}
\item{geometry}{sfc_MULTILINESTRING. The spatial geometry of the road segment.}
}
}
\source{
Provided by the municipal authorities of Curitiba (IPPUC).
}
\usage{
ippuc_road_axis
}
\description{
A dataset containing road information for Curitiba, Brazil.
}
\keyword{datasets}
