% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_calc.R
\name{nds_calc_dist}
\alias{nds_calc_dist}
\title{Calculate traveled distance from the naturalistic data}
\usage{
nds_calc_dist(data, geom, by, units = "meters")
}
\arguments{
\item{data}{A sf object with linestring geometry}

\item{geom}{The geometry column in the sf object}

\item{by}{A column name to group the results}

\item{units}{The desired distance unit ("meters", "kilometers")}
}
\value{
A tibble with traveled distance results
}
\description{
\code{nds_calc_dist} extracts traveled time from the naturalistic data.
}
\details{
This function uses a sf object (\link{nds_create_lines}) as main
input and extracts its traveled distance, grouping the results by the
attributes of the desired variable (parameter \code{by}). The distance
unit can be in "meters" (default) or "kilometers". The \code{geom}
parameter is used to identify the column that contains the geometry of the
sf object.
}
\examples{
path <- system.file("extdata", package = "ndsbr")
nds_data <- nds_load_data("driver", path)
nds_lines <- nds_create_lines(nds_data, x = LONG, y = LAT)
nds_calc_dist(nds_lines, wkt_lines, DRIVER)
}
\seealso{
\link{nds_create_lines}; \link{nds_calc_time}
}
