% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_calc.R
\name{nds_calc_speeding}
\alias{nds_calc_speeding}
\title{Calculate speeding rate from the naturalistic data}
\usage{
nds_calc_speeding(data, type, by, spd, exp, percentage = FALSE)
}
\arguments{
\item{data}{Object with naturalistic data (tibble or sf)}

\item{type}{Character with the type of speeding results
("time" or "distance")}

\item{by}{A column name to group the results}

\item{spd}{Numeric value of the speeding threshold}

\item{exp}{Numeric value of the exposure threshold}

\item{percentage}{Boolean value establishing if thresholds are percentages
(TRUE or FALSE)}
}
\value{
A tibble with speeding rate results
}
\description{
\code{nds_calc_distance} extracts the speeding rate from the
naturalistic data
}
\details{
This function uses the naturalistic data as main
input and extracts the speeding rate , grouping the results by the
attributes of the desired variable (parameter \code{by}). If \code{type} is
set to "distance", \code{data} must be a sf object with linestring geometry.
Otherwise, it can be a tibble or data.frame. \code{spd} and \code{exp} sets
two thresholds to identify exposure and speeding situations, based on the
performed speeds and speed limits. If \code{percentage} is TRUE, these
thresholds are set as percentages of the speed limit.
}
\examples{
path <- system.file("extdata", package = "ndsbr")
nds_data <- nds_load_data("driver", path)
nds_calc_speeding(nds_data, type = "time", by = ID, spd = 5, exp = 10)
}
