% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_calc.R
\name{nds_calc_time}
\alias{nds_calc_time}
\title{Calculate traveled time from the naturalistic data}
\usage{
nds_calc_time(data, by, units = "seconds", valid = "all")
}
\arguments{
\item{data}{A tibble or data.frame object containing naturalistic data}

\item{by}{A column name to group the results}

\item{units}{The desired time unit ("seconds", "minutes" or "hours")}

\item{valid}{Option to select all data or valid data ("all", "yes").}
}
\value{
A tibble with traveled time results.
}
\description{
\code{nds_calc_time} extracts traveled time from the naturalistic data.
}
\details{
This function uses a naturalistic data sample (\link{nds_load_data}) as main
input and extracts its traveled time, grouping the results by the attributes
of the desired variable (parameter \code{by}). The time unit can be in
"seconds" (default), "minutes" or "hours". It is possible to filter only
valid time data using the \code{valid} parameter. In it's default, the
function uses all data, including invalid times.
}
\examples{
path <- system.file("extdata", package = "ndsbr")
df <- nds_load_data("driver", path)
nds_calc_time(df, "DRIVER")
}
\seealso{
\link{nds_load_data}; \link{nds_calc_dist}
}
