% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_create.R
\name{nds_create_points}
\alias{nds_create_points}
\title{Create a sf object with point geometry}
\usage{
nds_create_points(data, x, y, valid = "all")
}
\arguments{
\item{data}{A data.frame or tibble object with coordinates.}

\item{x}{Data attribute with x coordinates.}

\item{y}{Data attribute with y coordinates.}

\item{valid}{Option to select all data or valid data ("all", "yes").}
}
\value{
A sf object with point geometry
}
\description{
This function creates a sf object using the point geometry, based on
the NDS-BR naturalistic data (see \link{nds_load_data}).
}
\details{
\code{nds_create_points} takes the naturalistic data as input and creates a
point spatial object. \code{x} and \code{y} arguments considers
longitude and latitude, respectively. It is possible to filter only valid
time data using the \code{valid} parameter. In it's default, the function
transforms all data, including invalid times.
}
\examples{
path <- system.file("extdata", package = "ndsbr")
nds_data <- nds_load_data("driver", path)
nds_create_points(nds_data, x = LONG, y = LAT)
}
\seealso{
\link{nds_create_lines}
}
