% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_download.R
\name{nds_download_sf}
\alias{nds_download_sf}
\title{Download spatial data}
\usage{
nds_download_sf(url, ...)
}
\arguments{
\item{url}{Download url.}

\item{...}{Parameters passed to `sf::st_read()`}
}
\value{
A \code{sf} object.
}
\description{
\code{nds_download_sf} downloads spatial data in \code{.zip} format and
extracts into a \code{sf} object.
}
\examples{
url <- "https://ippuc.org.br/geodownloads/SHAPES_SIRGAS/EIXO_RUA_SIRGAS.zip"
road <- nds_download_sf(url)
}
