% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_join.R
\name{nds_join_neigh}
\alias{nds_join_neigh}
\title{Join NDS data with neighborhood data}
\usage{
nds_join_neigh(ndsbr_data, neigh_data, vars = "NOME")
}
\arguments{
\item{ndsbr_data}{An `sf` object representing NDS data. This object should
have a `POINT` geometry type.}

\item{neigh_data}{An `sf` object representing neighborhood data. This object
should have a `POLYGON` geometry type.}

\item{vars}{A character vector of variables to join from the neighborhood
data. Default is `"NOME"`.}
}
\value{
An `sf` object with the joined data, including neighborhood
information.
}
\description{
This function joins NDS data with neighborhood data, ensuring the correct
spatial reference and geometry types.
}
\examples{
result <- nds_join_neigh(ndsbr_data_sf, ippuc_neigh)
}
