% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_load_data.R
\name{nds_load_data}
\alias{nds_load_data}
\title{Load naturalistic data stored in \code{.csv} files}
\usage{
nds_load_data(pattern, folder = NULL, sep = ";", vars = NULL)
}
\arguments{
\item{pattern}{A \code{character} object with a common initial text among
the files to be read.}

\item{folder}{A \code{character} object of the path to the folder containing
the files.}

\item{sep}{A \code{character} object defining the separator in the
\code{.csv} file. Default is ";".}

\item{vars}{A \code{character} vector of variable names to load. If empty,
all variables are loaded.}
}
\value{
A tibble of the NDS-BR dataset.
}
\description{
This function loads one or more \code{.csv} files containing the NDS-BR data.
}
\details{
\code{nds_load_data} default considers that data inside the \code{.csv} files
are separated by ';', and decimals are expressed by comma (','). If
\code{sep} is set to ",", decimals must be expressed by ".". The
\code{pattern} argument is used to identify a common initial text among the
files. If only one file is to be loaded, \code{pattern} can be the full name
of the file. In the \code{folder} argument it is possible to insert a path
of the folder in which the files are included. If this field is left empty,
the function considers that the files are present in the working directory
of the project (\code{getwd()}). In the \code{vars} parameter it is possible
to insert a character vector with the desired variable names to load.
}
\examples{
## Considering driver_A.csv, driver_B.csv and driver_C.csv:
path <- system.file("extdata", package = "ndsbr")
df <- nds_load_data("driver", path)
}
