% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_split_data.R
\name{nds_split_data}
\alias{nds_split_data}
\title{Split naturalistic data into the original files}
\usage{
nds_split_data(data, driver = .data$DRIVER)
}
\arguments{
\item{data}{A data.frame or tibble with naturalistic data}

\item{driver}{A column with the names of the drivers.}
}
\value{
A list with four tibbles of the NDS-BR dataset.
}
\description{
This function splits the naturalistic data into the original separation of
the NDS-BR sample.
}
\details{
\code{nds_split_data} takes the naturalistic data imported by
\link{nds_load_data} and splits into four tibbles, matching the original
separation of drivers provided by the sample of NDS-BR. The function uses
the column containing driver names to properly split the data.
}
\examples{
path <- system.file("extdata", package = "ndsbr")
df <- nds_load_data("driver", path)
df_split <- nds_split_data(df)
}
