% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsbr-package.R
\docType{package}
\name{ndsbr-package}
\alias{ndsbr}
\alias{ndsbr-package}
\title{ndsbr: Load, Analyze and Manipulate Data of The Brazilian Naturalistic Driving Study}
\description{
A compilation of methods used in the Brazilian Naturalistic Driving Study (NDS-BR). Contains functions to load, manipulate and analyze naturalistic data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pabsantos.github.io/ndsbr/}
}

}
\author{
\strong{Maintainer}: Pedro Augusto Borges dos Santos \email{pabsantos21@gmail.com}

}
\keyword{internal}
