% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsbr.R
\docType{package}
\name{ndsbr}
\alias{ndsbr}
\title{\code{ndsbr}: Load, Analyze and Manipulate Data From The Brazilian
Naturalistic Driving Study}
\description{
The \code{ndsbr} package provides two main categories of functions:
\code{create} and \code{calc}, in addition to other utility functions.
}
\section{\code{create} functions}{

The \code{create} functions are used to create spatial objects in \code{sf}
format, using the naturalistic data as input. These functions are:
\itemize{
\item \code{nds_create_points()}
\item \code{nds_create_lines()}
}
}

\section{\code{calc} functions}{

The \code{calc} functions can be used to extract basic information of the
naturalistic sample (travel time or travel distance) and safety-related
variables, such as speeding and use of mobile phone (UMP). These functions
are:
\itemize{
\item \code{nds_calc_dist()}
\item \code{nds_calc_time()}
\item \code{nds_calc_speeding()}
\item \code{nds_calc_ump()}
}
}

\section{Other functions}{

\itemize{
\item \link{nds_load_data}: Load naturalistic data.
\item \link{nds_download_sf}: Download and load spatial data.
}
}

