% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndsbr_data_sf}
\alias{ndsbr_data_sf}
\title{Example NDS-BR Spatial Dataset}
\format{
A data frame and `sf` object with 18,012 observations and 33 variables:
\describe{
  \item{DRIVER}{chr. The identifier for the driver.}
  \item{DAY}{chr. The date of the trip.}
  \item{DAY_CORRIGIDO}{chr. The corrected date of the trip.}
  \item{03:00:00}{chr. The timestamp}
  \item{TRIP}{dbl. The trip number.}
  \item{ID}{chr. The trip identifier.}
  \item{PR}{chr. The timestamp in GMT-3 timezone}
  \item{H}{dbl. The hour component of the timestamp.}
  \item{M}{dbl. The minute component of the timestamp.}
  \item{S}{dbl. The second component of the timestamp.}
  \item{TIME_ACUM}{dbl. The accumulated time of the trip in seconds.}
  \item{SPD_MPH}{dbl. The speed in miles per hour.}
  \item{SPD_KMH}{dbl. The speed in kilometers per hour.}
  \item{ACEL_MS2}{dbl. The acceleration in meters per second squared.}
  \item{HEADING}{dbl. The heading or direction of travel.}
  \item{ALTITUDE_FT}{dbl. The altitude in feet.}
  \item{VALID_TIME}{chr. Indicates whether the part of the trip is valid ("Yes" or "No").}
  \item{TIMESTAMP_GPS}{chr. Indicates whether the GPS timestamp is valid ("No").}
  \item{CPOOL}{chr. Indicates whether the trip is carpooling ("Not carpooling").}
  \item{CPOOLING_CHECKED}{chr. Indicates whether carpooling was checked ("Not carpooling").}
  \item{WSB}{chr. Indicates whether the driver was using seatbelt ("Yes").}
  \item{UMP_YN}{chr. Indicates whether the driver was using cellphone ("No").}
  \item{UMP}{chr. Indicates whether the driver was using cellphone ("0").}
  \item{PICK_UP}{chr. Indicates the pick-up cellphone quantity ("0").}
  \item{ACTION}{chr. Indicates the action taken during the usage of cellphone (NA).}
  \item{GPS_FILE}{chr. The GPS file identifier (NA).}
  \item{CIDADE}{chr. The city name ("Curitiba").}
  \item{BAIRRO}{chr. The neighborhood name ("JARDIM DAS AMÉRICAS").}
  \item{NOME_RUA}{chr. The street name ("NPI").}
  \item{HIERARQUIA_CWB}{chr. The road hierarchy according to the city of Curitiba ("NPI").}
  \item{HIERARQUIA_CTB}{chr. The road hierarchy according to the Brazilian Traffic Code ("NPI").}
  \item{LIMITE_VEL}{chr. The speed limit ("NPI").}
  \item{geometry}{POINT. The spatial geometry of the trip location.}
}
}
\source{
NDS-BR researchers.
}
\usage{
ndsbr_data_sf
}
\description{
A dataset containing trip information of the NDS-BR research project
}
\keyword{datasets}
