% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndsdrivers}
\alias{ndsdrivers}
\title{NDS-BR participants data}
\format{
## `ndsdrivers`
A tibble with 32 rows and 10 columns:
\describe{
  \item{DRIVER}{Driver alphabetic code}
  \item{SEX}{Driver sex}
  \item{AGE}{Age of the driver during the data collection}
  \item{DAYS}{Duration of the driver participation in days}
  \item{APP_DRIVER}{Boolean indicating if the driver participated as a app driver}
  \item{LICENSE_AGE}{Year in which the driver obtained its driver's license}
  \item{BRAND}{Brand of the driver's vehicle}
  \item{MODEL}{Model of the driver's vehicle}
  \item{MODEL_YEAR}{Model year of the driver's vehicle}
  \item{HP}{Horsepower of the driver's vehicle}
}
}
\usage{
ndsdrivers
}
\description{
A table with information about the drivers which participated in NDS-BR
}
\keyword{datasets}
