% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_join.R
\name{nds_join_axis}
\alias{nds_join_axis}
\title{Join NDS data with road axis data}
\usage{
nds_join_axis(
  ndsbr_data,
  road_axis,
  axis_vars = c("NMVIA", "SVIARIO", "HIERARQUIA")
)
}
\arguments{
\item{ndsbr_data}{An `sf` object representing NDS data. This object
should have a `POINT` geometry type.}

\item{road_axis}{An `sf` object representing road axis data. This object
should have a `LINESTRING` or `MULTILINESTRING` geometry type.}

\item{axis_vars}{A character vector of variables to join from the road axis
data. Default is `c("NMVIA", "SVIARIO", "HIERARQUIA")`.}
}
\value{
An `sf` object with the joined data.
}
\description{
This function joins NDS data with road axis data, including a 10-meter buffer
around axis data and arranging names and road hierarchy data.
}
\examples{
result <- nds_join_axis(ndsbr_data_sf, ippuc_road_axis)
}
