% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nds_join.R
\name{nds_join_spdlimit}
\alias{nds_join_spdlimit}
\title{Join NDS data with OSM speed limit data}
\usage{
nds_join_spdlimit(ndsbr_data, osm_data, vars = "maxspeed")
}
\arguments{
\item{ndsbr_data}{An `sf` object representing NDS data.
This object should have a `POINT` geometry type.}

\item{osm_data}{An `sf` object representing OSM data.
This object should have a `LINESTRING` or `MULTILINESTRING` geometry type.}

\item{vars}{A character vector of variables to join from the OSM data.
Default is `"maxspeed"`.}
}
\value{
An `sf` object with the joined data,
including speed limit information.
}
\description{
This function joins NDS data with OSM speed limit data,
including a 10-meter buffer around axis data and arranging speed limit
information.
}
\examples{
result <- nds_join_spdlimit(ndsbr_data_sf, osm_data)
}
