% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndsbr.R
\docType{package}
\name{ndsbr}
\alias{ndsbr-package}
\alias{ndsbr}
\title{\code{ndsbr}: Load, Analyze and Manipulate Data of The Brazilian
Naturalistic Driving Study}
\description{
The \code{ndsbr} package provides two main categories of functions:
\code{nds_create} and \code{nds_calc}, in addition to other utility
functions.
}
\section{\code{nds_create} functions}{

The \code{nds_create} functions are used to create spatial objects in
\code{sf} format, using the naturalistic data as input. These functions are:
\itemize{
\item \link{nds_create_points}
\item \link{nds_create_lines}
}
}

\section{\code{nds_calc} functions}{

The \code{nds_calc} functions can be used to extract basic information of the
naturalistic sample (traveled time or traveled distance) and safety-related
variables, such as speeding. These functions are:
\itemize{
\item \link{nds_calc_dist}
\item \link{nds_calc_time}
\item \link{nds_calc_speeding}
}
}

\section{Other functions}{

\itemize{
\item \link{nds_load_data}: Load naturalistic data.
\item \link{nds_download_sf}: Download and load spatial data.
}
}

\author{
\strong{Maintainer}: Pedro Augusto Borges dos Santos \email{pabsantos21@gmail.com}

}
