% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndsdrivers}
\alias{ndsdrivers}
\title{NDS-BR participants data}
\format{
A data frame with 40 rows and 11 variables:
\describe{
\item{driver_id}{\if{html}{\out{<chr>}} Unique identifier for each driver, e.g., "A", "B", "C".}
\item{sex}{\if{html}{\out{<chr>}} Gender of the driver, e.g., "F" for female, "M" for male.}
\item{age}{\if{html}{\out{<int>}} Age of the driver, e.g., 32, 39, 20.}
\item{license_type}{\if{html}{\out{<chr>}} Type of driver's license, e.g., "B", "AB", "D".}
\item{vehicle_age}{\if{html}{\out{<int>}} Year of the vehicle's manufacture, e.g., 2012, 2010, 2011.}
\item{app_driver}{\if{html}{\out{<lgl>}} Indicates if the driver is an app driver, e.g., TRUE or FALSE.}
\item{license_date}{\if{html}{\out{<date>}} Date when the driver was issued their license, e.g., 2009-12-17.}
\item{vehicle_brand}{\if{html}{\out{<chr>}} Brand of the vehicle, e.g., "Chevrolet", "Renault", "VW".}
\item{vehicle_model}{\if{html}{\out{<chr>}} Model of the vehicle, e.g., "Prisma 1.4L LT", "Scenic PRI 1616V".}
\item{vehicle_hp}{\if{html}{\out{<chr>}} Horsepower of the vehicle, extracted from strings like "5P/97CV".}
\item{vehicle_transmission}{\if{html}{\out{<chr>}} Type of vehicle transmission, e.g., "manual", "auto".}
}
}
\usage{
ndsdrivers
}
\description{
A table with information about the drivers which participated in NDS-BR
}
\keyword{datasets}
